HUD, Drops, Increasing difficulty (#13)
* HUD
* Drops (health, shields and missiles)
* drop money
* Increasing difficulty
* update changelog

+++ Assets/Scripts/Asteroid.cs
@@ -2,7 +2,7 @@
 
 public class Asteroid : MonoBehaviour
 {
-    public float points = 5f;
+    public int points = 5;
 
     private Health health;
     private GameState gameState;
